+++ src/ClientCore/Rendering/Sprites/SpriteSheetFactory.cs
@@ -66,8 +66,7 @@ namespace Sovereign.ClientCore.Rendering.Sprites
         /// <exception cref="SurfaceException">Thrown if the spritesheet file cannot be loaded.</exception>
         public SpriteSheet LoadSpriteSheet(SpriteSheetDefinition definition)
         {
-            Logger.DebugFormat("Loading spritesheet {0}: {1}", definition.SheetId,
-                definition.Filename);
+            Logger.DebugFormat("Loading spritesheet {0}", definition.Filename);
 
             /* Attempt to load the spritesheet surface. */
             var surface = LoadSurfaceForSpritesheet(definition);